requirements.txt
PyYAML
pyserial
onnxruntime
numpy
opencv-python-headless
requests
